Move registry/ddl.py to core/ddl.py.
No realy code changes here, just reorganization and import adjustments.

Previous guess (on DM-22508) that this code belonged in the registry
subpackage going forward was false; datastores and dimensions need it
to be able to declare what they need from Registry.

We're keeping the name instead of going back to schema.py (which will
be deleted shortly) because "schema" is too overloaded; sometimes it
means "namespace", sometimes it means table definitions.
