+++ python/lsst/daf/butler/core/__init__.py
@@ -10,6 +10,7 @@ from .config import *
 from .configSupport import LookupKey
 from .composites import *
 from .constraints import *
+from . import ddl
 from .datasets import *
 from .datastore import *
 from .dbAuth import *
