+++ python/lsst/daf/butler/registry/__init__.py
@@ -19,5 +19,4 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-from . import ddl
 from .registry import *
