+++ tests/test_postgresql.py
@@ -34,9 +34,10 @@ except ImportError:
 
 import sqlalchemy
 
+from lsst.daf.butler import ddl
 from lsst.daf.butler.core.registryConfig import RegistryConfig
 from lsst.daf.butler.registry.databases.postgresql import PostgresqlDatabase
-from lsst.daf.butler.registry import ddl, Registry
+from lsst.daf.butler.registry import Registry
 from lsst.daf.butler.registry.tests import DatabaseTests, RegistryTests
