+++ tests/test_sqlite.py
@@ -32,10 +32,11 @@ import sqlalchemy
 
 from lsst.sphgeom import ConvexPolygon, UnitVector3d
 
+from lsst.daf.butler import ddl
 from lsst.daf.butler.core.registryConfig import RegistryConfig
 from lsst.daf.butler.registry.databases.sqlite import SqliteDatabase
 from lsst.daf.butler.registry.tests import DatabaseTests, RegistryTests
-from lsst.daf.butler.registry import Registry, ddl
+from lsst.daf.butler.registry import Registry
 
 TESTDIR = os.path.abspath(os.path.dirname(__file__))
